+++ apps/docs/src/app/layout.tsx
@@ -1,11 +1,6 @@
 import AnalyticsProvider from '@/components/analytics-provider';
 import { GovieLink } from '@/components/navigation/custom-link';
-import {
-  Footer,
-  HeaderProps,
-  LoadMaterialSymbols,
-  Stack,
-} from '@ogcio/design-system-react';
+import { Footer, HeaderProps, Stack } from '@ogcio/design-system-react';
 import '@ogcio/design-system-react/styles.css';
 import '@ogcio/theme-govie/theme.css';
 import type { Metadata } from 'next';
@@ -102,9 +97,6 @@ export default function RootLayout({
 
   return (
     <html lang="en" suppressHydrationWarning className={lato.className}>
-      <head>
-        <LoadMaterialSymbols />
-      </head>
       <body
         className={`transition duration-500 h-screen flex flex-col bg-white`}
       >